processors: improvements to ollama.wasm for configuration and memory correctness
