+++ processors/ollama/ollama.go
@@ -5,7 +5,6 @@ package main
 import (
 	"time"
 
-	"github.com/wasmvision/wasmvision-sdk-go/config"
 	"github.com/wasmvision/wasmvision-sdk-go/datastore"
 	"github.com/wasmvision/wasmvision-sdk-go/http"
 	"github.com/wasmvision/wasmvision-sdk-go/logging"
@@ -33,14 +32,14 @@ func process(image mat.Mat) mat.Mat {
 		req := []byte(template)
 		tmpl := cm.ToList[[]byte](req)
 
-		data := http.PostImage(url+"/api/generate", "application/json", tmpl, "response", uint32(image))
+		data, _, isErr := http.PostImage(url+"/api/generate", "application/json", tmpl, "response", uint32(image)).Result()
 		switch {
-		case data.IsErr():
+		case isErr:
 			logging.Error("HTTP error")
-		case len(*data.OK()) > 0:
+		case len(data) > 0:
 			ps := datastore.NewProcessorStore(1)
-			ps.Set("captions", "caption", *data.OK())
-			logging.Info(*data.OK())
+			ps.Set("captions", "caption", data)
+			logging.Info(data)
 		default:
 			logging.Info("No result from ollama")
 		}
@@ -49,46 +48,4 @@ func process(image mat.Mat) mat.Mat {
 	}
 
 	return image
-}
-
-var (
-	url      string
-	model    string
-	template string
-)
-
-const defaultURL = "http://localhost:11434"
-const defaultModel = "llava"
-
-func loadConfig() {
-	if url == "" {
-		conf := config.GetConfig("url")
-		if conf.IsErr() || len(*conf.OK()) == 0 {
-			url = defaultURL
-		} else {
-			url = *conf.OK()
-		}
-
-		logging.Info("Using Ollama server at " + url)
-	}
-
-	if model == "" {
-		conf := config.GetConfig("model")
-		if conf.IsErr() || len(*conf.OK()) == 0 {
-			model = defaultModel
-		} else {
-			model = *conf.OK()
-		}
-
-		logging.Info("Using Ollama model " + model)
-	}
-
-	if template == "" {
-		template = `{
-			"model": "` + model + `",
-			"prompt":"Describe what is in this picture in highly complimentary terms using 6 words or less.",
-			"stream": false,
-			"images": ["%IMAGE%"]
-		  }`
-	}
 }
